sm8475-common: overlay: Enable WIFI 6GHz band support
set config_wifi6ghzSupport to true to return
6GHz band support for wifiManager.is6GHzBandSupported() call.
If this config is not set, framework query the list of
valid frequencies for 6GHz band from kernel/nl80211.
But this call will return valid frequencies only if
country code is set. So it is necessary to set config_wifi6ghzSupport
to true to return chipset support for 6GHz for cases where
country code is not set.

Bug: 193788596
Test: atest android.net.wifi.cts.WifiManagerTest
Change-Id: I68700fbed773aeeb53e70e6e6df9b3697f51b0d0
Signed-off-by: schqiushui <orochi9999@gmail.com>
Signed-off-by: Nishant Kumar <www.rajsonu13@gmail.com>